SummaryBot/requirements.txt
ctransformer
flask
flask-httpauth
werkzeug
torch
gunicorn
